refactor: 💡 dump and submit pages
