+++ src/features/dump/actions/add-contents.ts
@@ -1,11 +1,11 @@
 "use server";
 import "server-only";
+import { formatCreateContentsMessage } from "@/apis/line-notify/format-for-line";
 import { sendLineNotifyMessage } from "@/apis/line-notify/send-message";
 import { postContents } from "@/apis/prisma/fetch-contents";
 import { ERROR_MESSAGES, SUCCESS_MESSAGES } from "@/constants";
 import { auth } from "@/features/auth/lib/auth";
 import type { ContentsContext } from "@/features/dump/stores/contents-context";
-import { formatCreateContentsMessage } from "@/features/dump/utils/format-for-line";
 import { validateContents } from "@/features/dump/utils/validate-contents";
 import type { ActionState } from "./type";
