+++ src/features/submit/actions/change-contents-status.ts
@@ -1,12 +1,15 @@
 "use server";
 import "server-only";
 import { LineNotifyError } from "@/apis/line-notify/error";
+import { formatChangeStatusMessage } from "@/apis/line-notify/format-for-line";
 import { sendLineNotifyMessage } from "@/apis/line-notify/send-message";
-import { revertNewsStatus, updateNewsStatus } from "@/apis/prisma/fetch-news";
+import {
+	revertContentsStatus,
+	updateContentsStatus,
+} from "@/apis/prisma/fetch-contents";
 import { ERROR_MESSAGES } from "@/constants";
 import { auth } from "@/features/auth/lib/auth";
-import { formatChangeStatusMessage } from "@/features/dump/utils/format-for-line";
-import type { ServerAction } from "@/types/server-action";
+import type { ServerAction } from "@/types";
 import { Prisma } from "@prisma/client";
 
 type Change = "UPDATE" | "REVERT";
@@ -14,15 +17,15 @@ type Change = "UPDATE" | "REVERT";
 const handleStatusChange = async (changeType: Change) => {
 	switch (changeType) {
 		case "UPDATE":
-			return await updateNewsStatus();
+			return await updateContentsStatus();
 		case "REVERT":
-			return await revertNewsStatus();
+			return await revertContentsStatus();
 		default:
 			throw new Error(ERROR_MESSAGES.UNEXPECTED);
 	}
 };
 
-export async function changeNewsStatus(
+export async function changeContentsStatus(
 	changeType: Change,
 ): Promise<ServerAction> {
 	try {
@@ -31,12 +34,13 @@ export async function changeNewsStatus(
 
 		const message = formatChangeStatusMessage(
 			await handleStatusChange(changeType),
-			"NEWS",
+			"CONTENTS",
 		);
 		await sendLineNotifyMessage(message);
 		return { success: true, message };
 	} catch (error) {
 		if (error instanceof Prisma.PrismaClientKnownRequestError) {
+			console.error(error.message);
 			await sendLineNotifyMessage(error.message);
 			return {
 				success: false,
@@ -44,6 +48,7 @@ export async function changeNewsStatus(
 			};
 		}
 		if (error instanceof LineNotifyError) {
+			console.error(error.message);
 			// MEMO: unhandled防止 await sendLineNotifyMessage(error.message);
 			return {
 				success: false,
@@ -51,8 +56,10 @@ export async function changeNewsStatus(
 			};
 		}
 		if (error instanceof Error) {
+			console.error("Unexpected error:", error.message);
 			await sendLineNotifyMessage(error.message);
 		} else {
+			console.error("Unexpected error:", error);
 			await sendLineNotifyMessage(ERROR_MESSAGES.UNEXPECTED);
 		}
 		return {
